Code review fixes: svgFloat rounding, deterministic formatting, dedup constants
- Add svgFloat type with MarshalXMLAttr rounding to 2dp for cross-platform determinism
- Cast all float64 struct literal fields to svgFloat in edges/nodes/subgraphs/renderer
- Add String() method to diagramKind for readable error messages
- Export DefaultFontSize from renderer, deduplicate from output/svg
- Replace %g with %.2f in all style strings for predictable output
- Add doc comment to directionToRankDir about DirectionUnknown→TB default
- Regenerate golden files to match new formatting

+++ pkg/renderer/flowchart/renderer_test.go
@@ -71,10 +71,10 @@ func TestRenderRectangleNodeGeometry(t *testing.T) {
 	}
 	wantX := nl.X - nl.Width/2 + pad
 	wantY := nl.Y - nl.Height/2 + pad
-	if rect.X != wantX {
+	if float64(rect.X) != wantX {
 		t.Errorf("rect.X = %f, want %f", rect.X, wantX)
 	}
-	if rect.Y != wantY {
+	if float64(rect.Y) != wantY {
 		t.Errorf("rect.Y = %f, want %f", rect.Y, wantY)
 	}
